+++ js/modules/constant.js
@@ -1,0 +1,38 @@
+const AVATAR_COUNT = 6;
+const IMAGE_COUNT = 25;
+
+const LIKE_MAX_COUNT = 1000;
+const COMMENT_MAX_COUNT = 10;
+
+const NAMES = [
+  'Николай',
+  'Олег',
+  'Алиса',
+  'Антон',
+  'Галина',
+  'Матвей',
+  'Ника',
+  'Диана',
+  'Марина'
+];
+
+const DESCRIPTIONS = [
+  'Анапа 2010',
+  'Чемпионат Мира 2018',
+  'Я с мамой',
+  'Мой кот',
+  'На даче'
+];
+
+const MESSAGES = [
+  'Супер',
+  'ок',
+  'привет',
+  'норм',
+  'топчик',
+  '?',
+  'промокод сбермаркет скидка 500р на покупку от 1500: RPZ2NT7L687FJ'
+];
+
+
+export { AVATAR_COUNT, IMAGE_COUNT, LIKE_MAX_COUNT, COMMENT_MAX_COUNT, NAMES, DESCRIPTIONS, MESSAGES };
